InnovateX(C++) question 8

+++ C++/Easy/8.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    float x = 5.0;
+    int x = 5;
     if (x == 5)
     {
         cout << "x is 5" << endl;
